+++ R/frame.R
@@ -136,22 +136,57 @@ roboplotr_set_background <- function(p, color = getOption("roboplot.colors.backg
   p |> layout(paper_bgcolor = color, plot_bgcolor = color)
 }
 
+#' Zooming configuration
+#'
+#' Customize zooming behavior of [roboplots][roboplot()].
+#'
+#' @param zoom Characters. Controls how [roboplots][roboplot()] are zoomed in.
+#' Valid strings are "none", "scroll" (scrollwheel and two-finger drag) and "drag"
+#' (mouse or finger drag). Currently not used in [robomaps][robomap()]. Default is "scroll".
+#' @param ... Placeholder for future arguments.
+#' @export
+#' @examples
+#'
+#'
+#' # Zoom by dragging with mouse.
+#' energiantuonti |>
+#'   dplyr::filter(Alue == "Venäjä") |>
+#'   roboplot(Suunta, zoom = "drag")
+#'
+#' # Better for mobile to scroll with two fingers, or mobile on desktop.
+#' energiantuonti |>
+#'   dplyr::filter(Alue == "Venäjä") |>
+#'   roboplot(Suunta, zoom = "scroll")
+#'
+#' # Note that modebar zooming is not affected by this setting.
+#' energiantuonti |>
+#'   dplyr::filter(Alue == "Venäjä") |>
+#'   roboplot(Suunta, zoom = "none",
+#'            modebar = set_modebar(buttons = c("zoomin","zoomout","pan")))
+#'
+set_zoom <- function(zoom = getOption("roboplot.scroll"), ...) {
+  roboplotr_typecheck(zoom, "character", allow_null = FALSE)
+  roboplotr_valid_strings(zoom, c("none","drag","scroll"), any)
+
+  .res <- structure(zoom, class = c("roboplotr","roboplotr.set_zoom", class(zoom)))
+
+  .res
+}
+
 #' @importFrom plotly layout
 #' @importFrom stringr str_remove
-roboplotr_set_margin <-function(p, margin) {
+roboplotr_set_margin <-function(p, margin, zoom = "scroll") {
 
   if (!is.list(margin)) {
     stop("Insert plot margins as list (default is list(t,r,b,l,pad))\nNote that top and bottom margins will be currently ignored but programmatically set instead.", call. = F)
   } else if (any(!names(margin) %in% c("t","r","b","l","pad")) | any(!is.double(unlist(margin)))) {
     stop("All plot margins must be of double type, and named one or more of t, r, b, l or pad.\nNote that top and bottom margins will be currently ignored but programmatically set instead.", call. = F)
   } else {
-    # drag_mode <- getOption("roboplot.modebar.buttons")
-    drag_mode <- "zoom"#if (any(c("zoom","zoomin2d","pan") %in% drag_mode)) { str_remove(drag_mode[drag_mode %in% c("zoom","zoomin2d","pan")][1],"in2d") } else { "zoom" }
     p |>
       layout(
         margin = margin,
         autosize = T,
-        dragmode = drag_mode
+        dragmode = ifelse(zoom == "drag", T, F)
       )
     # p
   }