+++ R/zzz.R
@@ -46,6 +46,7 @@
     roboplot.shinyapp = FALSE,
     roboplot.xaxis.ceiling = "default",
     roboplot.zeroline = list(color = "black", width = 2),
+    roboplot.zoom = "scroll",
     roboplot.verbose = "All",
     roboplot.cur.options = "roboplotr"
   )